Support externalized ArrayBuffer for node::Buffer
