+++ filenames.gypi
@@ -326,6 +326,8 @@
       'atom/renderer/atom_renderer_client.h',
       'atom/renderer/guest_view_container.cc',
       'atom/renderer/guest_view_container.h',
+      'atom/renderer/node_array_buffer_bridge.cc',
+      'atom/renderer/node_array_buffer_bridge.h',
       'atom/utility/atom_content_utility_client.cc',
       'atom/utility/atom_content_utility_client.h',
       'chromium_src/chrome/browser/browser_process.cc',
